+++ src/app/engines/dependencies.engine.ts
@@ -1,6 +1,9 @@
 import * as _ from 'lodash';
 
-export class DependenciesEngine {
+import { finderInAngularAPIs } from '../../utils/angular-api';
+
+class DependenciesEngine {
+    private static _instance:DependenciesEngine = new DependenciesEngine();
     rawData: Object;
     modules: Object[];
     components: Object[];
@@ -10,7 +13,17 @@ export class DependenciesEngine {
     routes: Object[];
     pipes: Object[];
     classes: Object[];
-    constructor(data: Object) {
+    constructor() {
+        if(DependenciesEngine._instance){
+            throw new Error('Error: Instantiation failed: Use DependenciesEngine.getInstance() instead of new.');
+        }
+        DependenciesEngine._instance = this;
+    }
+    public static getInstance():DependenciesEngine
+    {
+        return DependenciesEngine._instance;
+    }
+    init(data: Object) {
         this.rawData = data;
         this.modules = _.sortBy(this.rawData.modules, ['name']);
         this.components = _.sortBy(this.rawData.components, ['name']);
@@ -20,6 +33,33 @@ export class DependenciesEngine {
         this.routes = _.sortBy(_.uniqWith(this.rawData.routes, _.isEqual), ['name']);
         this.pipes = _.sortBy(this.rawData.pipes, ['name']);
         this.classes = _.sortBy(this.rawData.classes, ['name']);
+    }
+    find(type: string) {
+        let finderInCompodocDependencies = function(data) {
+            let _result = {
+                    source: 'internal',
+                    data: null
+                },
+                i = 0,
+                len = data.length;
+            for (i; i<len; i++) {
+                if (type.indexOf(data[i].name) !== -1) {
+                    _result.data = data[i]
+                }
+            }
+            return _result;
+        },
+            resultInCompodocInjectables = finderInCompodocDependencies(this.injectables),
+            resultInCompodocClasses = finderInCompodocDependencies(this.classes),
+            resultInAngularAPIs = finderInAngularAPIs(type)
+
+        if (resultInCompodocInjectables.data !== null) {
+            return resultInCompodocInjectables
+        } else if (resultInCompodocClasses.data !== null) {
+            return resultInCompodocClasses
+        } else if (resultInAngularAPIs.data !== null) {
+            return resultInAngularAPIs
+        }
     }
     getModules() {
         return this.modules;
@@ -46,3 +86,5 @@ export class DependenciesEngine {
         return this.classes;
     }
 };
+
+export const $dependenciesEngine = DependenciesEngine.getInstance();